refactor(parser): move formal parameter checks from semantic to parser
Move the required-after-optional parameter check and TS signature
duplicate parameter name check out of the semantic checker and into
the parser, where the FormalParameters node is constructed. These
checks only read the node, so no new parser state is needed.

The duplicate-name check uses an allocation-free O(n²) scan instead of
a hash map; signature parameter lists are short and this is off the JS
hot path.
