+++ crates/oxc_semantic/src/checker/mod.rs
@@ -92,10 +92,6 @@ pub fn check<'a>(kind: AstKind<'a>, ctx: &SemanticBuilder<'a>) {
         }
         AstKind::Super(sup) => js::check_super(sup, ctx),
 
-        AstKind::FormalParameters(params) => {
-            ts::check_formal_parameters(params, ctx);
-        }
-
         AstKind::AwaitExpression(expr) => js::check_await_expression(expr, ctx),
         AstKind::ObjectExpression(expr) => js::check_object_expression(expr, ctx),
         AstKind::UnaryExpression(expr) => js::check_unary_expression(expr, ctx),
