+++ crates/oxc_semantic/src/checker/typescript.rs
@@ -1,12 +1,9 @@
 use std::borrow::Cow;
 
 use itertools::Itertools;
-use rustc_hash::FxHashMap;
 
 use oxc_ast::{AstKind, ast::*};
-use oxc_ecmascript::BoundNames;
 use oxc_span::{GetSpan, Span};
-use oxc_str::Str;
 
 use crate::{builder::SemanticBuilder, diagnostics};
 
@@ -74,32 +71,6 @@ pub fn check_ts_infer_type<'a>(infer_type: &TSInferType<'a>, ctx: &SemanticBuild
     if !is_in_conditional_extends_clause {
         ctx.error(diagnostics::infer_declaration_only_permitted_in_extends_clause(infer_type.span));
     }
-}
-
-pub fn check_formal_parameters(params: &FormalParameters, ctx: &SemanticBuilder<'_>) {
-    if params.kind == FormalParameterKind::Signature && params.items.len() > 1 {
-        check_duplicate_bound_names(params, ctx);
-    }
-
-    let mut has_optional = false;
-
-    for param in &params.items {
-        // function a(optional?: number, required: number) { }
-        if param.optional {
-            has_optional = true;
-        } else if has_optional && param.initializer.is_none() {
-            ctx.error(diagnostics::required_parameter_after_optional_parameter(param.span));
-        }
-    }
-}
-
-fn check_duplicate_bound_names<'a, T: BoundNames<'a>>(bound_names: &T, ctx: &SemanticBuilder<'_>) {
-    let mut idents: FxHashMap<Str<'a>, Span> = FxHashMap::default();
-    bound_names.bound_names(&mut |ident| {
-        if let Some(old_span) = idents.insert(ident.name.into(), ident.span) {
-            ctx.error(diagnostics::redeclaration(&ident.name, old_span, ident.span));
-        }
-    });
 }
 
 pub fn check_ts_module_declaration<'a>(decl: &TSModuleDeclaration<'a>, ctx: &SemanticBuilder<'a>) {
